+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - updated tests for set
 - updated tests for isExist
+- added test for randomElementFromSlice
 
 ## 0.2.1 - 2021.08.12